Revert unlicensed contribution
Some explanation is in order:

I have tried to contact Zanculmarktum three times in the last
25 days, but I have not received any response.
Meanwhile, others are waiting to be able to use flake-compat, so
I have decided that it is best to move forward with the licensing
effort, considering that everyone else except Eelco has already
agreed, and quickly as well.

I have no hard feelings towards Zanculmarktum and I really do not
like having to make this commit.

This reverts commit e363cffac26b0009972e0e20c1bbb0953473bd86 and
subsequent improvements.

I should note that it could be argued that their contribution was
not a creative work in the sense of most copyright law, but I'd
prefer to avoid any confusion by choosing to reimplement their
contribution in a "clean room" manner.