feat: allow optional moving to nix store
According to https://github.com/NixOS/nix/issues/3121#issuecomment-1050510805, flake-compat should be that issue's workaround.

However, it tries to move src to the nix store always. That is a sane default, as it imitates closer what flakes do. However, it does not let this tool to be used comfortably as a workaround for that issue.

With this change, the default behaviour is preserved. However, the user can pass `moveSrcToStore = false` to avoid it.

@moduon MT-83